Add support for Quote operations with an AIK.
This commit adds support for Quote2, LoadKey2, and GetPubKey. The test code for
these functions depends on the existence of a file aikblob that contains an AIK
for the current TPM. Support for creating AIKs with this library will be added
later.

This commit also adds response authentication checking for replies from the TPM.