Added support for closing key handles using defer.
This commit adds a CloseKey method that can be deferred on a tpm.Handle to close
the key at the end of a function. This is useful for callers that execute
LoadKey2 and later need to clean up the state of the loaded key.

This commit also has some minor renaming inside one function.