Add support for taking ownership.
This commit adds support for TPM_TakeOwnership. This command only works when the
platform owner has asserted physical presence, so the test for this case is
skipped in general. The TestOwnerClear test case is also skipped, since it
forces the TPM into a state that can only be recovered by asserting physical
presence.