Update tests to use general auth values.
This commit adds support to the tpm test code for general auth values computed
as the SHA1 hash of a string or as the well-known auth value. It also fixes the
Quote and Quote2 tests to use the AIK auth value for quotes rather than the SRK
value. This wasn't caught in previous testing because the SRK and AIK auth
values were previously both the well-known secret of all zeroes.