Use io.ReadWriter to contact the TPM.
This changes the TPM function interface to take an io.ReadWriter instead
of a *os.File. Existing calls to the functions will still work, since
*os.File implements io.ReadWriter.

This change allows the code to communicate with, e.g., a TPM emulator
that uses a Unix domain socket. See the examples and the new OpenTPM
function for how this would work.