Fix a packing bug in command verification.
PR #6 pulled encoding out into the tpmutil package. It also changed the
interface to take variadic arguments instead of a slice of interface{}
elements. However, tests against a TPM emulator revealed that the verify
function was failing, and it turned out to be because the params were
not being expanded when they were passed to tpmutil.Pack. And
tpmutil.Pack refuses to pack slices that are not RawBytes or []byte. So,
verify was failing.

This change expands params in verify while passing them to tpmutil.Pack.

This also demonstrates that this code needs better regression tests.