Allow nil slices to be used with the TBS library
The existing code will panic if any nil slices are passed into the
tpmutil/tbs library. This is especially bad as GetTCGLog has useful
behavior if a zero length buffer is passed.

This change allows nil slices to be passed for all tbs methods. It also
tests that we can get the TCG log by either
  - passing in a progressivly larger buffer
  - passing in nil to get the exact size