ECC capability check (#46)
* Implement GetCapability for algorithm descriptions.

This required making GetCapability more generic at the expense of type
safety.

* Skip ECC tests on non-ECC capable TPM2s.

Unfortunately Intel PTT does not necessarily support ECC in its TPM2
implementation. On these TPMs the existing test suite fails. With this
change the capabilities of the TPMs are verified before the ECC testing
functions are invoked and skipped if ECC is unsupported.