Fix `decodeCertify` always returning a truncated signature in case of `ECDSA` signature.
This is mainly due to the fact that the old code called `tpmutil.UnpackBuf` on only one element that is `signature`.
This worked for RSA signatures as their corresponding structure `TPMS_SIGNATURE_RSASSA` contains only 1 field `TPM2B_PUBLIC_KEY_RSA`,
in addition to `TPMI_ALG_HASH` field.
However, it failed for ECDSA signatures because their corresponding structure is `TPMS_SIGNATURE_ECDSA`,
which contains 2 fields R and S - instead of just 1 - which are both `TPM2B_ECC_PARAMETER`, in addition to `TPMI_ALG_HASH` field.

Therefore, this can be fixed by calling `tpmutil.UnpackBuf` on 2 elements, that is R followed by S, which results in retrieving the whole signature.

Signed-off-by: El Mostafa IDRASSI <mostafa.idrassi@tutanota.com>