quote: Remove internal usage of ownerPassword with Quote
Quote only takes one authorization area, so our functions shouldn't take
two passwords. To avoid a breaking change, the function signature of
`Quote` and `QuoteRaw` are note changed, but `encodeQuote` drops the
parameter.

We should remove the superflous param from the public functions when we
next make a breaking change.

Signed-off-by: Joe Richey <joerichey@google.com>