rename tpmdirect to tpm2 and rename existing tpm2 as legacy/tpm2 (#304)
This change moves the existing tpm2 package out to legacy/tpm2 and renames the direct package as tpm2.

The KDFa and KDFe helpers from the legacy API (which are needed by the new API) are moved into the new API, with the legacy API calling into the new API for htem.

A small adapter layer is added at tpm2/adapter.go to provide the API that go-tpm-tools/simulator requires. In the future, we can remove it after go-tpm-tools/simulator has been updated to use the new API.

We may flatten the new tpm2 library's sub-libraries structures and commands together into tpm2 in a future change, perhaps as part of resolving #302.