flatten the new API to just tpm2 (#308)
Merge all of the old TPMDirect libraries together into tpm2. That includes helpers, KDFs, etc., as well as the commands and structures. This was necessary to avoid dependency cycles, and it probably makes for an easier developer experience anyway.

This change adds a couple exceptions to the linting tool:

* tpm2.TPM2BXyz unfortunately stutters a bit, that's the tradeoff
* Startup_ and Shutdown_ end with an underscore until we can get a release of go-tpm-tools with a simulator that links to the legacy API at its new path

Fixes #302