execute commands by value and tweak the generic interface
This change incorporates some PR feedback that suggested making Command
a generic interface and deleting the need for both command and response
structures to have boilerplate methods that just return the command
code. Now only commands have that.

This had the side effect of making marshalling in general take inputs
by-value instead of by-pointer, which is a little more clear anyway.