Introduce helper for formatting PCR selection bitmasks (#345)
* Introduce helper for formatting PCR selection bitmasks

This change introduces PCClientCompatible.PCRs(), a function that
converts a variadic list of PCR indices (as ints) into a PCR selection
bitmask. Because of the vagaries of TPM:

1. That the minimum size of a PCR selection bitmask is not 0, but
   related to the minimum number of PCRs specified by the profile
2. That the PC Client Platform TPM Profile specification mandates a
   minimum but not a maximum number of implementation PCR,

this change creates an interface that could be implemented for other TPM
profiles that specify different amounts of PCRs. The vast majority of
on-market TPMs will just work with PCClientCompatible.PCRs, even if they
implement more than 24 PCRs.

PCRs() can panic if given invalid values; this is to allow it to be
inlined into the definition of a structure that needs a PCR selection.

* Don't publish the interface for the PCR selection formatter just yet.

* use uint instead