Implement TCP TPM protocol
This change implements the basic TCP TPM protocol, which is useful for
connecting to a running TPM simulator. It also introduces some tests
which are skipped if the --cmd_addr or --plat_addr flags aren't
provided.

A number of TCP commands aren't implemented and left for a future
change.

Tested against a recent build of the TCG TPM simulator from
https://github.com/trustedcomputinggroup/tpm

use the closeTPM helper, caught by linter

Co-authored-by: anamendes23 <anamendes@google.com>