Check for padding bytes in EK cert
Some TPM chips "pad" the EK cert that is stored in the TPM to completely
fill up the NV index. Most TPM chips do NOT do this. If we try to parse
the certificate with this extra padding, the `x509.ParseCertificate`
function will return an "trailing bytes" error.

Before trying to parse the certificate, we need to identify how many
bytes should be stripped from bytes that were read from the TPM. To do
this, we can use the ASN.1 `Unmarshal` function to get the "padding"
bytes. From there, we can use the length of the padding bytes and remove
those bytes from the bytes we send to the `x509.ParseCertificate`
function so it will parse correctly.